requirements.txt
asgiref==3.7.2
Django==5.0.1
django-allauth==0.61.1
djangorestframework==3.14.0
google-ai-generativelanguage==0.6.4
google-api-core==2.19.0
google-api-python-client==2.132.0
google-auth==2.30.0
google-auth-httplib2==0.2.0
google-cloud==0.34.0
google-cloud-vision==3.7.2
google-generativeai==0.6.0
google-pasta==0.2.0
googleapis-common-protos==1.63.0
imageio==2.34.0
imageio-ffmpeg==0.5.1
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
regex==2023.12.25
moviepy
gunicorn
uvicorn
